Incorporate filter visualization into master
